fixed react warnings and errors.
tanstack router is nice and all, but the modelling of a top navigation menu that behaves differently based on the active child-route is a huge mess.
I think I will move the state management to zustand (again) at one point. But duplicated state is also crap. Anyway - YOLO
The cleanest way would be to remove the outlet magic from the root-route and assemble the components myself. That way the active route can collect the stuff for the top menu bar and the relationship between routes would be more clear. Hopefully...